Add files via upload
The goal of this project is to align two DNA sequences of any length by identifying the optimal match through the strategic insertion of gaps in each strand to maximize alignment. For instance, to align the sequences "CAT" and "TCAT," a gap ("-") is added to form "-CAT," resulting in an edit distance of 2. The edit distance represents the minimum cost required to achieve the best possible alignment. For more complex sequences, a 2D cost matrix is calculated to determine the minimal alignment cost and generate the corresponding alignment strings. This process accounts for the higher penalty associated with adding gaps. The approach is specifically designed to efficiently handle extremely long DNA sequences.